feat(service): implement all method from brand category support to product service

+++ products/src/main/java/com/oralie/products/sevice/impl/BrandServiceImpl.java
@@ -3,11 +3,15 @@ package com.oralie.products.sevice.impl;
 
 import com.oralie.products.dto.request.BrandRequest;
 import com.oralie.products.dto.response.BrandResponse;
+import com.oralie.products.dto.response.ListResponse;
+import com.oralie.products.exception.ResourceAlreadyExistException;
 import com.oralie.products.exception.ResourceNotFoundException;
 import com.oralie.products.model.Brand;
 import com.oralie.products.repository.BrandRepository;
 import com.oralie.products.sevice.BrandService;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.sql.results.graph.basic.BasicResultGraphNode;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -23,13 +27,20 @@ public class BrandServiceImpl implements BrandService {
     private final BrandRepository brandRepository;
 
     @Override
-    public List<BrandResponse> getAllBrands(int page, int size, String sortBy, String sort) {
+    public ListResponse<BrandResponse> getAllBrands(int page, int size, String sortBy, String sort) {
         Sort sortObj = sort.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(sortBy).ascending() : Sort.by(sortBy).descending();
         Pageable pageable = PageRequest.of(page, size, sortObj);
+        Page<Brand> pageBrands = brandRepository.findAll(pageable);
+        List<Brand> brands = pageBrands.getContent();
 
-        List<Brand> brands = brandRepository.findAll(pageable).getContent();
-
-        return mapToBrandResponseList(brands);
+        return ListResponse.<BrandResponse>builder()
+                .data(mapToBrandResponseList(brands))
+                .pageNo(pageBrands.getNumber())
+                .pageSize(pageBrands.getSize())
+                .totalElements((int) pageBrands.getTotalElements())
+                .totalPages(pageBrands.getTotalPages())
+                .isLast(pageBrands.isLast())
+                .build();
     }
 
     @Override
@@ -39,6 +50,8 @@ public class BrandServiceImpl implements BrandService {
 
     @Override
     public BrandResponse createBrand(BrandRequest brandRequest) {
+        if (brandRepository.existsByName(brandRequest.getName()))
+            throw new ResourceAlreadyExistException("Brand already exists with name " + brandRequest.getName());
         Brand brand = Brand.builder()
                 .name(brandRequest.getName())
                 .description(brandRequest.getDescription())
@@ -52,6 +65,8 @@ public class BrandServiceImpl implements BrandService {
     @Override
     public BrandResponse updateBrand(Long id, BrandRequest brandRequest) {
         Brand brand = brandRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Brand not found", "id", id + ""));
+        if (brandRepository.existsByName(brandRequest.getName()))
+            throw new ResourceAlreadyExistException("Brand already exists with name " + brandRequest.getName());
         brand.setName(brandRequest.getName());
         brand.setDescription(brandRequest.getDescription());
         brand.setImageUrl(brandRequest.getUrlImage());
@@ -63,7 +78,11 @@ public class BrandServiceImpl implements BrandService {
     @Override
     public void deleteBrand(Long id) {
         Brand brand = brandRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Brand not found", "id", id + ""));
-        brandRepository.delete(brand);
+        if (brandRepository.existsByIdAndProductsIsEmpty(id)) {
+            brandRepository.delete(brand);
+        } else {
+            throw new ResourceNotFoundException("Brand cannot be deleted as it contains products", "id", id + "");
+        }
     }
 
     private BrandResponse mapToBrandResponse(Brand brand) {
